Revert "Feature/storepart/dto"
