+++ src/main/java/pinup/backend/store/entity/Inventory.java
@@ -1,0 +1,28 @@
+package pinup.backend.store.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import pinup.backend.member.command.domain.Users;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Inventory {
+
+    @EmbeddedId
+    private InventoryKey id;
+
+    @Column(name = "earned_at", nullable = false)
+    private LocalDateTime earnedAt;
+
+    @Column(name = "is_equipped", nullable = false)
+    private boolean isEquipped = true;
+
+}
