+++ src/main/java/pinup/backend/store/entity/Store.java
@@ -1,4 +1,4 @@
-package pinup.backend.store.domain;
+package pinup.backend.store.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -6,6 +6,8 @@ import lombok.*;
 @Entity
 @Getter
 @NoArgsConstructor
+@AllArgsConstructor
+@Builder
 @Table(name = "store")    //store로 이름 변경시 필요한지 피드백 필요
 public class Store {
 
@@ -15,10 +17,10 @@ public class Store {
     @Column(name = "item_id")
     private int itemId;
 
-    //행정구역ID = FK(관계있음) 한지역에 아이템 여러개있음
-    //@ManyToOne
-    //@JoinColumn(name = "region_code", nullable = true)
-    //private Region region;
+    // 행정구역ID = FK(관계있음) 한지역에 아이템 여러개있음
+    // @ManyToOne
+    //@JoinColumn(name = "reion_code", nullable = true)
+    //private region region;
 
     // 아이템 이름
     @Column(nullable = false, length = 50)
@@ -33,19 +35,19 @@ public class Store {
     private int price;
 
     // 아이템 카테고리 (Enum)
-    @Enumerated(EnumType.STRING)
+    @Enumerated
     @Column(nullable = false, length = 20)
     private StoreItemCategory category;
 
-    // 아이템 이미지 URL
+    // 아이템 이미지 경로
     @Column(name = "image_url", nullable = false, length = 255)
     private String imageUrl;
 
-    // 아이템 판매 여부
+    // 첫 생성시 판매 중
     @Column(name = "is_active", nullable = false)
     private boolean isActive = true;
 
-    // 아이템 판매 중지
+    // 판매중지 메소드
     public void deactivate() {
         this.isActive = false;
     }
